refactor: more reorganizing
